+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Commands `build list`, `build delete` and `build run`
 - [server] Support for teresa.yaml v2 format
 - Command `service whitelist-source-ranges`
+- Firewall whitelist to the `service info` command
 
 ### Fixed
 - [server] Fix panic when monitoring the rolling update